Create module for user-specific configuration
This module encapsulates anything related to user-specific configuration
(as opposed to system configuration). The idea is that if this module is
disabled, then user-specific stuff will be omitted during the system
rebuild.

Right now, this module just configures a user (me!), sets up the Home
Manager module, and re-exports the `home.file` option for symlinking
files into the home directory iff `modules.home.enable` is `true`.