Install Visual Studio Code
The `settings.json` file in the Nix store is symlinked into the VSCode
config directory. Since the Nix store is read-only, user settings are
immutable, which is super cool. But now any changes to them will involve
a rebuild of the entire system, which is not super cool.

It's definitely not the best workflow for iterating quickly, but as a
workaround, I found that I can use workflow settings as an overlay for
dirty changes.

I am debating to myself whether or not if I should symlink dotfiles into
the files in this git repository instead of the Nix store. Considering
maybe using something like `mkOutOfStoreSymlink` for this.